pool: use uintptr for pool size
note, arenas and blocks still use ulong for sizes. so
we have to check for overflow when attempting to merge
arenas.

+++ sys/include/pool.h
@@ -1,11 +1,11 @@
 typedef struct Pool Pool;
 struct Pool {
 	char*	name;
-	ulong	maxsize;
+	uintptr	maxsize;
 
-	ulong	cursize;
-	ulong	curfree;
-	ulong	curalloc;
+	uintptr	cursize;
+	uintptr	curfree;
+	uintptr	curalloc;
 
 	ulong	minarena;	/* smallest size of new arena */
 	ulong	quantum;	/* allocated blocks should be multiple of */
